pyproject.toml
[project]
name = "bytegenie-final-assignment"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Huvinesh Rajendran", email = "huvinesh@teleme.co" }
]
dependencies = [
    "polars>=1.5.0",
    "matplotlib>=3.9.2",
    "networkx>=3.3",
    "psycopg2>=2.9.9",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
    "scipy>=1.14.1",
    "pandas>=2.2.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bytegenie_final_assignment"]
